requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.9
pydantic-settings==2.11.0
aiofiles==24.1.0
transformers==4.55.0
torch==2.8.0
Pillow==11.3.0
accelerate==1.10.1
openai-whisper==20250625
numpy==2.3.2
